fix(agents): satisfy continuation delivery checks
